refactoring: Переименовал пакет, согласно замечанию.
