+++ src/main/java/ru/practicum/shareit/request/dto/ItemRequestDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.requests.dto;
+package ru.practicum.shareit.request.dto;
 
 /**
  * // TODO .
