+++ src/main/java/ru/practicum/shareit/request/mapper/ItemRequestsMapper.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.request.mapper;
+
+public class ItemRequestsMapper {
+}
