+++ src/main/java/ru/practicum/shareit/request/model/ItemRequest.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.requests.model;
+package ru.practicum.shareit.request.model;
 
 /**
  * // TODO .
