--- src/main/java/ru/practicum/shareit/requests/mapper/ItemRequestsMapper.java
@@ -1,4 +1,0 @@
-package ru.practicum.shareit.requests.mapper;
-
-public class ItemRequestsMapper {
-}
